Unify the implementation of "mmap/munmap" in different allocators to avoid duplication.
