+++ src/server/memory/malloc.cc
@@ -28,20 +28,42 @@
 
 #include "server/memory/malloc.h"
 
+#include <sys/mman.h>
+
 #include <stddef.h>
 #include <string>
 #include <vector>
 
+#include "common/util/flags.h"
 #include "common/util/logging.h"
 
 namespace vineyard {
 
 namespace memory {
 
+/// Gap between two consecutive mmap regions allocated by fake_mmap.
+/// This ensures that the segments of memory returned by
+/// fake_mmap are never contiguous and dlmalloc does not coalesce it
+/// (in the client we cannot guarantee that these mmaps are contiguous).
+constexpr int64_t kMmapRegionsGap = sizeof(size_t);
+
+// Fine-grained control for whether we need pre-populate the shared memory.
+//
+// Usually it causes a long wait time at the start up, but it could improved
+// the performance of visiting shared memory.
+//
+// In cases that the startup time doesn't much matter, e.g., in kubernetes
+// environment, pre-populate will archive a win.
+DEFINE_bool(reserve_memory, false, "Pre-reserving enough memory pages");
+
 std::unordered_map<void*, MmapRecord> mmap_records;
 
 static void* pointer_advance(void* p, ptrdiff_t n) {
   return (unsigned char*) p + n;
+}
+
+static void* pointer_retreat(void* p, ptrdiff_t n) {
+  return (unsigned char*) p - n;
 }
 
 static ptrdiff_t pointer_distance(void const* pfrom, void const* pto) {
@@ -88,6 +110,64 @@ int create_buffer(int64_t size) {
   }
 #endif
   return fd;
+}
+
+void* mmap_buffer(int64_t size, bool* is_committed, bool* is_zero) {
+  // Add kMmapRegionsGap so that the returned pointer is deliberately not
+  // page-aligned. This ensures that the segments of memory returned by
+  // fake_mmap are never contiguous.
+  size += kMmapRegionsGap;
+
+  int fd = create_buffer(size);
+  CHECK_GE(fd, 0) << "Failed to create buffer during mmap";
+  // MAP_POPULATE can be used to pre-populate the page tables for this memory
+  // region
+  // which avoids work when accessing the pages later. However it causes long
+  // pauses
+  // when mmapping the files. Only supported on Linux.
+
+  int mmap_flag = MAP_SHARED;
+  if (FLAGS_reserve_memory) {
+#ifdef __linux__
+    mmap_flag |= MAP_POPULATE;
+    *is_committed = true;
+#endif
+  }
+
+  void* pointer = mmap(NULL, size, PROT_READ | PROT_WRITE, mmap_flag, fd, 0);
+  if (pointer == MAP_FAILED) {
+    LOG(ERROR) << "mmap failed with error: " << strerror(errno);
+    return pointer;
+  }
+
+  MmapRecord& record = mmap_records[pointer];
+  record.fd = fd;
+  record.size = size;
+
+  // We lie to dlmalloc/mimalloc about where mapped memory actually lives.
+  pointer = pointer_advance(pointer, kMmapRegionsGap);
+  return pointer;
+}
+
+int munmap_buffer(void* addr, int64_t size) {
+  addr = pointer_retreat(addr, kMmapRegionsGap);
+  size += kMmapRegionsGap;
+
+  auto entry = mmap_records.find(addr);
+
+  if (entry == mmap_records.end() || entry->second.size != size) {
+    // Reject requests to munmap that don't directly match previous
+    // calls to mmap, to prevent dlmalloc from trimming.
+    return -1;
+  }
+
+  int r = munmap(addr, size);
+  if (r == 0) {
+    close(entry->second.fd);
+  }
+
+  mmap_records.erase(entry);
+  return r;
 }
 
 void GetMallocMapinfo(void* addr, int* fd, int64_t* map_size,
