+++ src/server/memory/mimalloc.cc
@@ -15,8 +15,6 @@ limitations under the License.
 
 #include "server/memory/mimalloc.h"
 
-#include <sys/mman.h>
-
 #include <stdio.h>
 
 #include "common/util/status.h"
@@ -30,24 +28,13 @@ namespace memory {
 
 void* MimallocAllocator::Init(const size_t size) {
   // create memory using mmap
-  int fd = create_buffer(size);
-  void* space = mmap(NULL, size, PROT_READ | PROT_WRITE, MAP_SHARED, fd, 0);
+  bool is_committed = false;
+  bool is_zero = true;
+  void* space = mmap_buffer(size, &is_committed, &is_zero);
   if (space == nullptr) {
     return space;
   }
-
-  // the addr must be 64MB aligned(required by mimalloc)
-  // this will cause some loss of memory, and the maximum does not exceed 64MB
-  void* addr = reinterpret_cast<void*>(
-      (reinterpret_cast<uintptr_t>(space) + MIMALLOC_SEGMENT_ALIGNED_SIZE - 1) &
-      ~(MIMALLOC_SEGMENT_ALIGNED_SIZE - 1));
-  size_t real_size = size - ((size_t) addr - (size_t) space);
-
-  MmapRecord& record = mmap_records[space];
-  record.fd = fd;
-  record.size = size;
-
-  return Mimalloc::Init(addr, real_size);
+  return Mimalloc::Init(space, size, is_committed, is_zero);
 }
 
 }  // namespace memory
